Refactored and Structured a Little More
